Reorganize utils tests
