+++ tests/utils/test_memory.py
@@ -1,0 +1,11 @@
+"""Tests for memory utilities."""
+
+
+from baybe.utils.memory import bytes_to_human_readable
+
+
+def test_memory_human_readable_conversion():
+    """The memory conversion to human readable format is correct."""
+    assert bytes_to_human_readable(1024) == (1.0, "KB")
+    assert bytes_to_human_readable(1024**2) == (1.0, "MB")
+    assert bytes_to_human_readable(4.3 * 1024**4) == (4.3, "TB")
