+++ tests/utils/test_numerical.py
@@ -1,0 +1,28 @@
+"""Tests for numerical utilities."""
+import numpy as np
+import pytest
+from pytest import param
+
+from baybe.utils.numerical import closest_element
+
+_TARGET = 1337
+_CLOSEST = _TARGET + 0.1
+
+
+@pytest.mark.parametrize(
+    "as_ndarray", [param(False, id="list"), param(True, id="array")]
+)
+@pytest.mark.parametrize(
+    "array",
+    [
+        param(_CLOSEST, id="0D"),
+        param([0, _CLOSEST], id="1D"),
+        param([[2, 3], [0, _CLOSEST]], id="2D"),
+    ],
+)
+def test_closest_element(as_ndarray, array):
+    """The closest element can be found irrespective of the input type."""
+    if as_ndarray:
+        array = np.asarray(array)
+    actual = closest_element(array, _TARGET)
+    assert actual == _CLOSEST, (actual, _CLOSEST)
